Refactor code to fix state reloading when re routing
