+++ src/components/pages/Rockets/Rockets.js
@@ -1,30 +1,23 @@
 /* eslint linebreak-style: ["error", "windows"] */
 
-import React, { useEffect, useRef } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React from 'react';
+import { useDispatch } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
 import {
   Container, Row, Col, Image, Button,
 } from 'react-bootstrap';
-import { buttonToggle, fetchRockets } from '../../../redux/rockets/rockets';
+import PropTypes from 'prop-types';
+import { buttonToggle } from '../../../redux/rockets/rockets';
 import './Rocket.css';
 
-const Rockets = () => {
+const Rockets = (props) => {
   const dispatch = useDispatch();
-  const rockets = useSelector((state) => state.rocketsReducer);
-  const dataFetchedRef = useRef(false);
-  useEffect(() => {
-    if (dataFetchedRef.current);
-    else {
-      dataFetchedRef.current = true;
-      dispatch(fetchRockets());
-    }
-  }, [rockets]);
+  const { rockets } = props;
   return (
     <Container fluid className="container">
       {rockets.map((element) => (
         <Row key={uuidv4()}>
-          <Col className="eachbox" xs={5}><Image fluid src={element.flickr_images[0]} alt={element.flickr_images[0]} /></Col>
+          <Col className="eachbox" xs={5}><Image fluid src={element.flickr_images} alt="Flicker Image" /></Col>
           <Col className="eachbox">
             <h1>{element.name}</h1>
             <p>{element.description}</p>
@@ -34,6 +27,17 @@ const Rockets = () => {
       ))}
     </Container>
   );
+};
+
+Rockets.propTypes = {
+  rockets: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    type: PropTypes.string.isRequired,
+    flickr_images: PropTypes.string.isRequired,
+    description: PropTypes.string.isRequired,
+    reserved: PropTypes.bool.isRequired,
+  })).isRequired,
 };
 
 export default Rockets;
